fix: fix fetch performance issue and refactor fetching & protected modal state
